Fixed Motion issues, the player moves smoothly

+++ Artifacts/src/Vector2f.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 struct Vector2f {
-    float x, y;
-    Vector2f(float x = 0.0f, float y = 0.0f)
+    int x, y;
+    Vector2f(int x = 0.0f, int y = 0.0f)
     {
         this->x = x;
         this->y = y;
